version2.1.3 数据库迁移

+++ rems_platform/src/main/java/com/tan/dto/DeviceProperties.java
@@ -22,7 +22,7 @@ public class DeviceProperties {
     /**
      * 血氧
      */
-    private Integer bloodOxygen;
+    private Double bloodOxygen;
 
 
 }
